+++ wled00/cfg.cpp
@@ -1,4 +1,5 @@
 #include "wled.h"
+#include "wled_ethernet.h"
 
 /*
  * Serializes and parses the cfg.json and wsec.json settings files, stored in internal FS.
@@ -106,10 +107,11 @@ bool deserializeConfig(JsonObject doc, bool fromFS) {
 
 
   JsonArray ins = hw_led["ins"];
+
+  uint16_t lC = 0;
+
   if (fromFS || !ins.isNull()) {
-    uint8_t s = 0; //bus iterator
-    strip.isRgbw = false;
-    strip.isOffRefreshRequred = false;
+    uint8_t s = 0;  // bus iterator
     busses.removeAll();
     uint32_t mem = 0;
     for (JsonObject elm : ins) {
@@ -128,22 +130,21 @@ bool deserializeConfig(JsonObject doc, bool fromFS) {
       uint8_t colorOrder = (int)elm[F("order")];
       uint8_t skipFirst = elm[F("skip")];
       uint16_t start = elm["start"] | 0;
+      if (length==0 || start + length > MAX_LEDS) continue; // zero length or we reached max. number of LEDs, just stop
       uint8_t ledType = elm["type"] | TYPE_WS2812_RGB;
       bool reversed = elm["rev"];
-
+      bool refresh = elm["ref"] | false;
+      ledType |= refresh << 7;  // hack bit 7 to indicate strip requires off refresh
+      s++;
+      uint16_t busEnd = start + length;
+      if (busEnd > lC) lC = busEnd;
       BusConfig bc = BusConfig(ledType, pins, start, length, colorOrder, reversed, skipFirst);
-      if (bc.adjustBounds(ledCount)) {
-        //RGBW mode is enabled if at least one of the strips is RGBW
-        strip.isRgbw = (strip.isRgbw || BusManager::isRgbw(ledType));
-        //refresh is required to remain off if at least one of the strips requires the refresh.
-        strip.isOffRefreshRequred |= BusManager::isOffRefreshRequred(ledType);
-        s++;
-        mem += busses.memUsage(bc);
-        if (mem <= MAX_LED_MEMORY) busses.add(bc);
-      }
+      mem += BusManager::memUsage(bc);
+      if (mem <= MAX_LED_MEMORY && busses.getNumBusses() <= WLED_MAX_BUSSES) busses.add(bc);  // finalization will be done in WLED::beginStrip()
     }
-    strip.finalizeInit(ledCount);
+    // finalization done in beginStrip()
   }
+  if (lC > ledCount) ledCount = lC; // fix incorrect total length (honour analog setup)
   if (hw_led["rev"]) busses.getBus(0)->reversed = true; //set 0.11 global reversed setting for first bus
 
   // read multiple button configuration
@@ -321,6 +322,7 @@ bool deserializeConfig(JsonObject doc, bool fromFS) {
   JsonObject if_live = interfaces["live"];
   CJSON(receiveDirect, if_live["en"]);
   CJSON(e131Port, if_live["port"]); // 5568
+  if (e131Port == DDP_DEFAULT_PORT) e131Port = E131_DEFAULT_PORT; // prevent double DDP port allocation
   CJSON(e131Multicast, if_live[F("mc")]);
 
   JsonObject if_live_dmx = if_live[F("dmx")];
@@ -576,6 +578,25 @@ void serializeConfig() {
   #ifdef WLED_USE_ETHERNET
   JsonObject ethernet = doc.createNestedObject("eth");
   ethernet["type"] = ethernetType;
+  if (ethernetType != WLED_ETH_NONE && ethernetType < WLED_NUM_ETH_TYPES) {
+    JsonArray pins = ethernet.createNestedArray("pin");
+    for (uint8_t p=0; p<WLED_ETH_RSVD_PINS_COUNT; p++) pins.add(esp32_nonconfigurable_ethernet_pins[p].pin);
+    if (ethernetBoards[ethernetType].eth_power>=0)     pins.add(ethernetBoards[ethernetType].eth_power);
+    if (ethernetBoards[ethernetType].eth_mdc>=0)       pins.add(ethernetBoards[ethernetType].eth_mdc);
+    if (ethernetBoards[ethernetType].eth_mdio>=0)      pins.add(ethernetBoards[ethernetType].eth_mdio);
+    switch (ethernetBoards[ethernetType].eth_clk_mode) {
+      case ETH_CLOCK_GPIO0_IN:
+      case ETH_CLOCK_GPIO0_OUT:
+        pins.add(0);
+        break;
+      case ETH_CLOCK_GPIO16_OUT:
+        pins.add(16);
+        break;
+      case ETH_CLOCK_GPIO17_OUT:
+        pins.add(17);
+        break;
+    }
+  }
   #endif
 
   JsonObject hw = doc.createNestedObject("hw");
@@ -618,7 +639,9 @@ void serializeConfig() {
     ins[F("order")] = bus->getColorOrder();
     ins["rev"] = bus->reversed;
     ins[F("skip")] = bus->skippedLeds();
-    ins["type"] = bus->getType();
+    ins["type"] = bus->getType() & 0x7F;;
+    ins["ref"] = bus->isOffRefreshRequired();
+    ins[F("rgbw")] = bus->isRgbw();
   }
 
   // button(s)
@@ -643,7 +666,7 @@ void serializeConfig() {
 
   JsonObject hw_ir = hw.createNestedObject("ir");
   hw_ir["pin"] = irPin;
-  hw_ir[F("type")] = irEnabled;              // the byte 'irEnabled' does contain the IR-Remote Type ( 0=disabled )
+  hw_ir["type"] = irEnabled;  // the byte 'irEnabled' does contain the IR-Remote Type ( 0=disabled )
 
   JsonObject hw_relay = hw.createNestedObject(F("relay"));
   hw_relay["pin"] = rlyPin;