+++ wled00/data/index.js
@@ -1475,14 +1475,16 @@ var plJson = {"0":{
   "end": 0
 }};
 
-var plSelContent = "";
-function makePlSel(arr) {
-  plSelContent = "";
-  for (var i = 0; i < arr.length; i++) {
-    var n = arr[i][1].n ? arr[i][1].n : "Preset " + arr[i][0];
-    if (arr[i][1].playlist && arr[i][1].playlist.ps) continue; // remove playlists, sub-playlists not yet supported
-    plSelContent += `<option value=${arr[i][0]}>${n}</option>`
-  }
+function makePlSel(incPl=false) {
+	var plSelContent = "";
+	delete pJson["0"];	// remove filler preset
+	var arr = Object.entries(pJson);
+	for (var i = 0; i < arr.length; i++) {
+		var n = arr[i][1].n ? arr[i][1].n : "Preset " + arr[i][0];
+		if (!incPl && arr[i][1].playlist && arr[i][1].playlist.ps) continue; //remove playlists, sub-playlists not yet supported
+		plSelContent += `<option value=${arr[i][0]}>${n}</option>`
+	}
+	return plSelContent;
 }
 
 function refreshPlE(p) {
@@ -1569,8 +1571,8 @@ function makeP(i,pl) {
   <div class="c">Repeat <input class="noslide" type="number" id="pl${i}rp" oninput="plR(${i})" max=127 min=0 value=${rep>0?rep:1}> times</div>
   End preset:<br>
   <select class="btn sel sel-ple" id="pl${i}selEnd" onchange="plR(${i})" data-val=${plJson[i].end?plJson[i].end:0}>
-    <option value=0>None</option>
-    ${plSelContent}
+		<option value=0>None</option>
+    ${makePlSel(true)}
   </select>
   </div>
   <button class="btn btn-i btn-p" onclick="testPl(${i}, this)"><i class='icons btn-icon'>&#xe139;</i>Test</button>`;
@@ -1628,7 +1630,7 @@ function makePlEntry(p,i) {
   return `
   <div class="plentry">
     <select class="btn sel sel-pl" onchange="plePs(${p},${i},this)" data-val=${plJson[p].ps[i]} data-index=${i}>
-      ${plSelContent}
+		${makePlSel()}
     </select>
     <button class="btn btn-i btn-xs btn-pl-del" onclick="delPl(${p},${i})"><i class="icons btn-icon">&#xe037;</i></button>
     <div class="h plnl">Duration</div><div class="h plnl">Transition</div><div class="h pli">#${i+1}</div><br>